# Faster getModificationTimestamps routines