--- a/oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHNode.java
+++ b/oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHNode.java
@@ -12,8 +12,10 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
+import java.util.function.Predicate;
 import org.heigit.bigspatialdata.oshdb.osm.OSMEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMNode;
+import org.heigit.bigspatialdata.oshdb.osm.OSMRelation;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
 import org.heigit.bigspatialdata.oshdb.util.OSHDBBoundingBox;
 import org.heigit.bigspatialdata.oshdb.util.OSHDBTimestamp;
@@ -319,6 +321,24 @@ public class OSHNode extends OSHEntity<OSMNode> implements Iterable<OSMNode>, Se
     List<OSHDBTimestamp> result = new ArrayList<>(this.iterator().next().getVersion());
     for (OSMNode osmNode : this) {
       result.add(osmNode.getTimestamp());
+    }
+    return Lists.reverse(result);
+  }
+
+  @Override
+  public List<OSHDBTimestamp> getModificationTimestamps(Predicate<OSMEntity> osmEntityFilter) {
+    List<OSHDBTimestamp> result = new ArrayList<>(this.iterator().next().getVersion());
+    OSHDBTimestamp prevNonmatch = null;
+    for (OSMNode osmNode : this) {
+      if (osmNode.isVisible() && (osmEntityFilter == null || osmEntityFilter.test(osmNode))) {
+        if (prevNonmatch != null) {
+          result.add(prevNonmatch);
+          prevNonmatch = null;
+        }
+        result.add(osmNode.getTimestamp());
+      } else {
+        prevNonmatch = osmNode.getTimestamp();
+      }
     }
     return Lists.reverse(result);
   }