--- a/oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHWay.java
+++ b/oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHWay.java
@@ -22,6 +22,7 @@ import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeMap;
 import java.util.TreeSet;
+import java.util.function.Predicate;
 import org.heigit.bigspatialdata.oshdb.osm.OSMEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMMember;
 import org.heigit.bigspatialdata.oshdb.osm.OSMNode;
@@ -404,11 +405,33 @@ public class OSHWay extends OSHEntity<OSMWay> implements Serializable {
     }
   }
 
+
+  @Override
+  public List<OSHDBTimestamp> getModificationTimestamps(Predicate<OSMEntity> osmEntityFilter) {
+    return _getModificationTimestamps(true, osmEntityFilter);
+  }
+
   @Override
   public List<OSHDBTimestamp> getModificationTimestamps(boolean recurse) {
+    return _getModificationTimestamps(recurse, null);
+  }
+
+  private List<OSHDBTimestamp> _getModificationTimestamps(
+      boolean recurse,
+      Predicate<OSMEntity> osmEntityFilter
+  ) {
     List<OSHDBTimestamp> wayTs = new ArrayList<>(this.iterator().next().getVersion());
+    OSHDBTimestamp prevNonmatch = null;
     for (OSMWay osmWay : this) {
-      wayTs.add(osmWay.getTimestamp());
+      if (osmWay.isVisible() && (osmEntityFilter == null || osmEntityFilter.test(osmWay))) {
+        if (prevNonmatch != null) {
+          wayTs.add(prevNonmatch);
+          prevNonmatch = null;
+        }
+        wayTs.add(osmWay.getTimestamp());
+      } else {
+        prevNonmatch = osmWay.getTimestamp();
+      }
     }
     if (!recurse) {
       return Lists.reverse(wayTs);
@@ -416,11 +439,10 @@ public class OSHWay extends OSHEntity<OSMWay> implements Serializable {
 
     Map<OSHEntity, LinkedList<OSHDBTimestamp>> childEntityTs = new TreeMap<>();
 
-    int i = 0;
     OSHDBTimestamp nextT = new OSHDBTimestamp(Long.MAX_VALUE);
     for (OSMWay osmWay : this) {
-      OSHDBTimestamp thisT = wayTs.get(i++);
-      if (!osmWay.isVisible()) {
+      OSHDBTimestamp thisT = osmWay.getTimestamp();
+      if (!osmWay.isVisible() || (osmEntityFilter != null && !osmEntityFilter.test(osmWay))) {
         nextT = thisT;
         continue;
       }