--- a/oshdb/src/test/java/org/heigit/bigspatialdata/oshdb/osh/OSHRelationTest.java
+++ b/oshdb/src/test/java/org/heigit/bigspatialdata/oshdb/osh/OSHRelationTest.java
@@ -142,38 +142,40 @@ public class OSHRelationTest {
     OSHNode hnode1 = OSHNode.build(n1versions);
     List<OSMNode> n2versions = new ArrayList<>();
     n2versions.add(new OSMNode(124l, 4, new OSHDBTimestamp(12l), 24l, 0, new int[]{}, 0, 0));
-    n2versions.add(new OSMNode(124l, 3, new OSHDBTimestamp(8l), 23l, 0, new int[]{}, 0, 0));
+    n2versions.add(new OSMNode(124l, 3, new OSHDBTimestamp(9l), 23l, 0, new int[]{}, 0, 0));
     n2versions.add(new OSMNode(124l, 2, new OSHDBTimestamp(4l), 22l, 0, new int[]{}, 0, 0));
     n2versions.add(new OSMNode(124l, 1, new OSHDBTimestamp(3l), 21l, 0, new int[]{}, 0, 0));
     OSHNode hnode2 = OSHNode.build(n2versions);
     List<OSMNode> n3versions = new ArrayList<>();
-    n3versions.add(new OSMNode(125l, 4, new OSHDBTimestamp(11l), 34l, 0, new int[]{}, 0, 0));
-    n3versions.add(new OSMNode(125l, 3, new OSHDBTimestamp(9l), 33l, 0, new int[]{}, 0, 0));
+    n3versions.add(new OSMNode(125l, 3, new OSHDBTimestamp(11l), 34l, 0, new int[]{}, 0, 0));
     n3versions.add(new OSMNode(125l, 2, new OSHDBTimestamp(6l), 32l, 0, new int[]{}, 0, 0));
     n3versions.add(new OSMNode(125l, 1, new OSHDBTimestamp(1l), 31l, 0, new int[]{}, 0, 0));
     OSHNode hnode3 = OSHNode.build(n3versions);
 
     List<OSMWay> w1versions = new ArrayList<>();
-    w1versions.add(new OSMWay(1, 2, new OSHDBTimestamp(7l), 4445l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0)}));
-    w1versions.add(new OSMWay(1, 1, new OSHDBTimestamp(5l), 4444l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0), new OSMMember(125, OSMType.NODE, 0)}));
+    w1versions.add(new OSMWay(1, 3, new OSHDBTimestamp(7l), 4445l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0)}));
+    w1versions.add(new OSMWay(1, 2, new OSHDBTimestamp(5l), 4444l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0), new OSMMember(125, OSMType.NODE, 0)}));
+    w1versions.add(new OSMWay(1, 1, new OSHDBTimestamp(4l), 4443l, 23, new int[]{}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 0), new OSMMember(124, OSMType.NODE, 0), new OSMMember(125, OSMType.NODE, 0)}));
     OSHWay hway1 = OSHWay.build(w1versions, Arrays.asList(hnode1, hnode2, hnode3));
 
     List<OSMRelation> versions = new ArrayList<>();
-    versions.add(new OSMRelation(1, 3, new OSHDBTimestamp(10l), 10000l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
-    versions.add(new OSMRelation(1, 2, new OSHDBTimestamp(8l), 10000l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{}));
-    versions.add(new OSMRelation(1, 1, new OSHDBTimestamp(5l), 10000l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
+    versions.add(new OSMRelation(1, -4, new OSHDBTimestamp(20l), 10004l, 1, new int[]{}, new OSMMember[]{}));
+    versions.add(new OSMRelation(1, 3, new OSHDBTimestamp(10l), 10003l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
+    versions.add(new OSMRelation(1, 2, new OSHDBTimestamp(8l), 10002l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(123, OSMType.NODE, 1)}));
+    versions.add(new OSMRelation(1, 1, new OSHDBTimestamp(5l), 10001l, 1, new int[]{1, 1, 2, 2}, new OSMMember[]{new OSMMember(1, OSMType.WAY, 0)}));
     OSHRelation hrelation = OSHRelation.build(versions, Arrays.asList(hnode1, hnode2, hnode3), Arrays.asList(hway1));
 
     List<OSHDBTimestamp> tss = hrelation.getModificationTimestamps(false);
     assertNotNull(tss);
-    assertEquals(3, tss.size());
+    assertEquals(4, tss.size());
     assertEquals(5l, tss.get(0).getRawUnixTimestamp());
     assertEquals(8l, tss.get(1).getRawUnixTimestamp());
     assertEquals(10l, tss.get(2).getRawUnixTimestamp());
+    assertEquals(20l, tss.get(3).getRawUnixTimestamp());
 
     tss = hrelation.getModificationTimestamps(true);
     assertNotNull(tss);
-    assertEquals(6, tss.size());
+    assertEquals(7, tss.size());
     assertEquals(5l, tss.get(0).getRawUnixTimestamp());
     assertEquals(6l, tss.get(1).getRawUnixTimestamp());
     assertEquals(7l, tss.get(2).getRawUnixTimestamp());
@@ -182,6 +184,7 @@ public class OSHRelationTest {
     assertEquals(10l,  tss.get(4).getRawUnixTimestamp());
     // timestamp 11 has to be missing because at the time, the node wasn't part of the way member of the relation anymore
     assertEquals(12l, tss.get(5).getRawUnixTimestamp());
+    assertEquals(20l, tss.get(6).getRawUnixTimestamp());
   }
 
   static OSHNode buildHOSMNode(List<OSMNode> versions) {