Improves the way modification timestamps for OSHEntity objects are calculated:

* avoid nested loops – this did work OK for small entities like ways with a couple of versions, but not for large (route) relations that have thousands of members, have hundreds of versions and it's members also have dozens of versions each. Before complexity was ~`O(v*m*mv)`, now it should be more like `O(v+m*mv*f)`, where `f` is a small factor (larger than one) representing the amount of fragmentation of the membership time intervals of the particular relation (which is usually small).
* calculate filtered modification timestamps directly (without first calculating all modification timestamps and filtering them afterwards).
* test osh entities early with the given osm entity filter (e.g. when filtering for `highway=residential`, ways that have the `highway` tag but never had the value `residential` don't need to be looked at in further processing steps).

This resolves #8