@tyrasd 2018-09-03
Here's a quick performance test:

```
    OSHDBH2 oshdb = new OSHDBH2("…/rheinland-pfalz.oshdb");
    Integer result = OSMEntitySnapshotView
        .on(oshdb)
        .areaOfInterest(new OSHDBBoundingBox(7.458,49.016,8.236,49.58))
        .timestamps("2018-01-01")
        .osmType(OSMType.RELATION)
        .osmTag("type", "route")
        .osmEntityFilter(e -> e.getId() == 88376)
        .count();
```

This takes ~ 8 min on current master (and oshdb v0.4), and ~5 sec on this branch.

The normal benchmarks also seem to be slightly faster:

```
# PR

building histocounts: 0.96s ±0.41s
highway histocounts (length): 0.73s ±0.01s
forest histocounts (area): 5.60s ±0.05s
unique contributors (simple area of interest): 1.29s ±0.11s
count roads (complex area of interest): 3.50s ±0.06s
object history by id: 0.23s ±0.02s

# master

building histocounts: 1.28s ±0.55s
highway histocounts (length): 0.84s ±0.02s
forest histocounts (area): 7.03s ±0.66s
unique contributors (simple area of interest): 1.28s ±0.07s
count roads (complex area of interest): 3.67s ±0.10s
object history by id: 0.31s ±0.01s
```